requirements.txt
mysql-connector-python~=9.1.0
pandas~=2.2.3
numpy~=2.1.3
flask~=3.1.0
werkzeug~=3.1.3